# server/docs/openapi.yaml
openapi: 3.0.0
info:
  description: >-
    API for the CSC302 Cancer Care Ontario SDC form management and submission
    portal
  version: 1.0.0
  title: CSC302-Team 4 SDC App API Spec
...
paths:
...
  /response:
    post:
      tags:
        - Response
      operationId: startResponse
      summary: Start a new SDCFormResponse
      requestBody:
        description: The form response parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SDCFormResponse'
      responses:
        '200':
          description: 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SDCFormResponse'
              examples:
                newResponse:
                  value:
                    _id: 1439vsdj90320jf90m34
                    diagnosticProcedureID: some_procedure
                    patientID: 098
                    formFillerID: '123'
        '500':
          description: 500 response
    put:
      tags:
        - Response
      operationId: submitForm
      requestBody:
        description: The form response to submit
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SDCFormResponse'
      responses:
        '200':
          description: 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SDCPersistentLink'
              examples:
                submission:
                  value:
                    link: kasfndsk
        '500':
          description: 500 response
  '/response/{responseID}':
    get:
      tags:
        - Response
      operationId: getResponses
      summary: Get a specific SDCFormResponse
      parameters:
        - in: path
          name: responseID
          description: The ID of the SDCFormResponse
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 200 response
          content:
            application/pdf:
              schema:
                $ref: '#/components/schemas/SDCFormResponse'
        '404':
          description: 404 response
  /response/search:
    get:
      tags:
        - Response
      operationId: searchResponses
      summary: Search for an existing SDCFormResponse
      parameters:
        - in: query
          name: diagnosticProcedureID
          description: >-
            DiagnosticProcedureID of an SDCFormResponse as a persistent
            rendering
          required: false
          schema:
            type: string
        - in: query
          name: patientID
          description: PatientID of the specific SDCFormResponse
          required: false
          schema:
            type: string
        - in: query
          name: formFillerID
          description: FormFillerID of the specific SDCFormResponse
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SDCFormResponse'
        '404':
          description: 404 response
  /response/answer:
    post:
      tags:
        - Response
      operationId: submitAnswer
      summary: Submit an SDCFormAnswer for an SDCResponse
      requestBody:
        description: The SDCFormAnswer to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SDCFormAnswer'
      responses:
        '200':
          description: 200 response
        '500':
          description: 500 response
    delete:
      tags:
        - Response
      operationId: deleteAnswer
      summary: Delete an SDCFormAnswer from an SDCResponse
      parameters:
        - in: query
          name: responseID
          description: The ID of the SDCFormResponse to remove the answer from
          required: true
          schema:
            type: string
        - in: query
          name: nodeID
          description: >-
            The ID of the SDCFormNode the answer refers to. Used in conjuction
            with responseID to find answer to delete
          required: true
          schema:
            type: string
        - in: query
          name: choiceID
          description: >-
            The id of the SDCFormChoice if we are deleting a specific choice and
            not the entire answer
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 200 response
        '404':
          description: 404 response
        '500':
          description: 500 response
...
components:
  schemas:
...
    SDCFormResponse:
      type: object
      properties:
        diagnosticProcedureID:
          type: string
          description: ID of the procedure this response corresponds to
        formVersion:
          type: integer
          description: Which version of the form was filled out
        patientID:
          type: string
          description: The patient this form was filled out for
        formFillerID:
          type: string
          description: The clinitian who filled out this form
        answers:
          type: array
          items:
            $ref: '#/components/schemas/SDCFormAnswer'
        persistentLinks:
          type: array
          items:
            type: string
    SDCPersistentLink:
      type: object
      properties:
        link:
          type: string
        response:
          $ref: '#/components/schemas/SDCFormResponse'
    SDCFormAnswer:
      type: object
      properties:
        responseID:
          type: string
        nodeID:
          type: string
        field:
          $ref: '#/components/schemas/SDCFieldAnswer'
        choices:
          type: array
          items:
            $ref: '#/components/schemas/SDCChoiceAnswer'
    SDCFieldAnswer:
      type: object
      properties:
        stringValue:
          type: string
        numberValue:
          oneOf:
            - type: number
            - type: integer
    SDCChoiceAnswer:
      type: object
      properties:
        choiceID:
          type: string
        field:
          $ref: '#/components/schemas/SDCFieldAnswer'
servers:
  - url: 'localhost:3001/api'
    variables: {}